+++ modules/dark-whispers.mjs
@@ -10,7 +10,7 @@ export default class DarkWhispers {
           let response = ""
           // eslint-disable-next-line capitalized-comments
           // data-action tells us what button was clicked
-          switch ($(event.currentTarget).attr("data-action")) {
+          switch (event.currentTarget.dataset.action) {
             case "accept":
               response = `${game.i18n.format("GMTOOLKIT.Message.DarkWhispers.Accepted", { currentUser: actor.name })}`
               // Adjusting Corruption is left as a manual intervention.
@@ -23,7 +23,7 @@ export default class DarkWhispers {
           }
 
           // Add the ask from the original message
-          response += `<blockquote>${$(event.currentTarget).attr("data-ask")}</blockquote>`
+          response += `<blockquote>${event.currentTarget.dataset.ask}</blockquote>`
           let chatData = {
             speaker: ChatMessage.getSpeaker("token"),
             content: response,